Fix conflicting function prototype for rng_fill_buffer on 32-bit architectures
In rng.h, rng_fill_buffer is declared with `size_t size`, whereas rng.c
implemented it with `uint64_t size`.

On 32-bit architectures (such as armhf or i386), `size_t` is 32-bit while
`uint64_t` is 64-bit, leading to a conflicting types compiler error when
building rng.c.

Align the implementation signature in rng.c to use `size_t` for the buffer
size and loop index.

Tested-by: 林博仁(Buo-ren Lin) <buo.ren.lin@gmail.com>

+++ rng.c
@@ -27,9 +27,9 @@ int32_t rng_get_random_number(device_testing_context_type *device_testing_contex
     return result;
 }
 
-void rng_fill_buffer(device_testing_context_type *device_testing_context, char *buffer, uint64_t size) {
+void rng_fill_buffer(device_testing_context_type *device_testing_context, char *buffer, size_t size) {
     int32_t *int_buffer = (int32_t *) buffer;
-    uint64_t i;
+    size_t i;
     for(i = 0; i < (size / 4); i++) {
         int_buffer[i] = rng_get_random_number(device_testing_context);
     }
